EVM-891: Refactor fuzz for reuse.
