+++ ipld/hamt/fuzz/fuzz_targets/simple.rs
@@ -2,51 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 #![no_main]
-use arbitrary::Arbitrary;
-use fvm_ipld_hamt::Hamt;
+use fvm_ipld_hamt::Config;
 use libfuzzer_sys::fuzz_target;
 
-#[derive(Debug, Arbitrary)]
-struct Operation {
-    key: u64,
-    method: Method,
-}
+mod common;
 
-#[derive(Debug, Arbitrary)]
-enum Method {
-    Insert(u64),
-    Remove,
-    Get,
-}
-
-fuzz_target!(|data: (u8, Vec<Operation>)| {
+fuzz_target!(|data: (u8, Vec<common::Operation>)| {
     let (flush_rate, operations) = data;
-    let db = fvm_ipld_blockstore::MemoryBlockstore::default();
-    let mut hamt = Hamt::<_, _, _>::new_with_bit_width(&db, 5);
-    let mut elements = ahash::AHashMap::new();
-
-    let flush_rate = (flush_rate as usize).saturating_add(5);
-    for (i, Operation { key, method }) in operations.into_iter().enumerate() {
-        if i % flush_rate == 0 {
-            // Periodic flushing of Hamt to fuzz blockstore usage also
-            hamt.flush().unwrap();
-        }
-
-        match method {
-            Method::Insert(v) => {
-                elements.insert(key, v);
-                hamt.set(key, v).unwrap();
-            }
-            Method::Remove => {
-                let el = elements.remove(&key);
-                let hamt_deleted = hamt.delete(&key).unwrap().map(|(_, v)| v);
-                assert_eq!(hamt_deleted, el);
-            }
-            Method::Get => {
-                let ev = elements.get(&key);
-                let av = hamt.get(&key).unwrap();
-                assert_eq!(av, ev);
-            }
-        }
-    }
+    let mut conf = Config::default();
+    conf.bit_width = 5;
+    common::run(flush_rate, operations, conf);
 });
